requirements.txt
neo4j==5.20.0
google-generativeai==0.4.1
numpy==1.26.4
python-dotenv==1.0.1
